requirements.txt
streamlit~=1.46.0
openai
langchain
langchain-openai~=0.3.25
protobuf~=5.29.5
python-dotenv~=1.1.1